Move Kotlin tests into the new harness
