--- packages/pigeon/platform_tests/android_kotlin_unit_tests/README.md
@@ -1,3 +1,0 @@
-# android_kotlin_unit_tests
-
-Unit-tests for Pigeon generated Kotlin code.  See [../../tools/run_tests.dart](../../tools/run_tests.dart).
