--- packages/pigeon/platform_tests/android_kotlin_unit_tests/android/gradle.properties
@@ -1,3 +1,0 @@
-org.gradle.jvmargs=-Xmx1536M
-android.useAndroidX=true
-android.enableJetifier=true
