--- packages/pigeon/platform_tests/android_kotlin_unit_tests/pubspec.yaml
@@ -1,20 +1,0 @@
-name: android_kotlin_unit_tests
-description: Unit tests for Pigeon generated Java code.
-
-publish_to: "none" # Remove this line if you wish to publish to pub.dev
-version: 1.0.0+1
-
-environment:
-  sdk: ">=2.12.0 <3.0.0"
-
-dependencies:
-  cupertino_icons: ^1.0.2
-  flutter:
-    sdk: flutter
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
