+++ packages/pigeon/platform_tests/test_plugin/android/build.gradle
@@ -43,4 +43,21 @@ android {
     defaultConfig {
         minSdkVersion 16
     }
+
+    testOptions {
+        unitTests.includeAndroidResources = true
+        unitTests.returnDefaultValues = true
+        unitTests.all {
+            testLogging {
+               events "passed", "skipped", "failed", "standardOut", "standardError"
+               outputs.upToDateWhen {false}
+               showStandardStreams = true
+            }
+        }
+    }
+
+    dependencies {
+        testImplementation 'junit:junit:4.+'
+        testImplementation "io.mockk:mockk:1.12.4"
+    }
 }
