+++ packages/pigeon/platform_tests/test_plugin/android/src/test/kotlin/com/example/test_plugin/EnumTest.kt
@@ -2,36 +2,36 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package com.example.android_kotlin_unit_tests
+package com.example.test_plugin
 
 import io.flutter.plugin.common.BinaryMessenger
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
+import io.mockk.verify
 import junit.framework.TestCase
 import org.junit.Test
 import java.nio.ByteBuffer
 import java.util.ArrayList
 
-class MultipleArityTests: TestCase() {
+internal class EnumTest: TestCase() {
     @Test
-    fun testSimpleHost() {
+    fun testEchoHost() {
         val binaryMessenger = mockk<BinaryMessenger>()
-        val api = mockk<MultipleArityHostApi>()
+        val api = mockk<EnumApi2Host>()
 
-        val inputX = 10L
-        val inputY = 5L
+        val channelName = "dev.flutter.pigeon.EnumApi2Host.echo"
+        val input = DataWithEnum(EnumState.SUCCESS)
 
-        val channelName = "dev.flutter.pigeon.MultipleArityHostApi.subtract"
         val handlerSlot = slot<BinaryMessenger.BinaryMessageHandler>()
 
         every { binaryMessenger.setMessageHandler(channelName, capture(handlerSlot)) } returns Unit
-        every { api.subtract(any(), any()) } answers { firstArg<Long>() - secondArg<Long>() }
+        every { api.echo(any()) } returnsArgument 0
 
-        MultipleArityHostApi.setUp(binaryMessenger, api)
+        EnumApi2Host.setUp(binaryMessenger, api)
 
-        val codec = MultipleArityHostApi.codec
-        val message = codec.encodeMessage(listOf(inputX, inputY))
+        val codec = EnumApi2Host.codec
+        val message = codec.encodeMessage(listOf(input))
         message?.rewind()
         handlerSlot.captured.onMessage(message) {
             it?.rewind()
@@ -39,36 +39,37 @@ class MultipleArityTests: TestCase() {
             val wrapped = codec.decodeMessage(it) as HashMap<String, Any>?
             assertNotNull(wrapped)
             wrapped?.let {
-                assertEquals(inputX - inputY, wrapped["result"])
+                assertTrue(wrapped.containsKey("result"))
+                assertEquals(input, wrapped["result"])
             }
         }
+
+        verify { binaryMessenger.setMessageHandler(channelName, handlerSlot.captured) }
+        verify { api.echo(input) }
     }
 
     @Test
-    fun testSimpleFlutter() {
+    fun testEchoFlutter() {
         val binaryMessenger = mockk<BinaryMessenger>()
-        val api = MultipleArityFlutterApi(binaryMessenger)
+        val api = EnumApi2Flutter(binaryMessenger)
 
-        val inputX = 10L
-        val inputY = 5L
+        val input = DataWithEnum(EnumState.SUCCESS)
 
         every { binaryMessenger.send(any(), any(), any()) } answers {
-            val codec = MultipleArityFlutterApi.codec
+            val codec = EnumApi2Flutter.codec
             val message = arg<ByteBuffer>(1)
             val reply = arg<BinaryMessenger.BinaryReply>(2)
             message.position(0)
             val args = codec.decodeMessage(message) as ArrayList<*>
-            val argX = args[0] as Long
-            val argY = args[1] as Long
-            val replyData = codec.encodeMessage(argX - argY)
+            val replyData = codec.encodeMessage(args[0])
             replyData?.position(0)
             reply.reply(replyData)
         }
 
         var didCall = false
-        api.subtract(inputX, inputY) {
+        api.echo(input) {
             didCall = true
-            assertEquals(inputX - inputY, it)
+            assertEquals(input, it)
         }
 
         assertTrue(didCall)
