+++ packages/pigeon/platform_tests/test_plugin/android/src/test/kotlin/com/example/test_plugin/PrimitiveTest.kt
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package com.example.android_kotlin_unit_tests
+package com.example.test_plugin
 
 import io.flutter.plugin.common.BinaryMessenger
 import io.mockk.every
